Springified reporting service
