+++ repose-aggregator/core/core-lib/src/main/java/com/rackspace/papi/service/context/impl/PowerApiContextManager.java
@@ -90,7 +90,6 @@ public class PowerApiContextManager implements ServletContextListener {
       PapiBanner.print(LOG);
       ca.getContext(ResponseMessageServiceContext.class).contextInitialized(sce);
       // TODO:Refactor - This service should be bound to a fitler-chain specific JNDI context
-      ca.getContext(ReportingServiceContext.class).contextInitialized(sce);
       ca.getContext(RequestHeaderServiceContext.class).contextInitialized(sce);
       ca.getContext(ResponseHeaderServiceContext.class).contextInitialized(sce);
       ca.getContext(DistributedDatastoreServiceContext.class).contextInitialized(sce);
