+++ repose-aggregator/core/valve/src/main/java/com/rackspace/cloud/valve/controller/service/context/impl/ReposeValveControllerContextManager.java
@@ -5,7 +5,6 @@ import com.rackspace.papi.service.context.ContextAdapter;
 import com.rackspace.papi.service.context.ServiceContext;
 import com.rackspace.papi.service.context.ServletContextHelper;
 import com.rackspace.papi.service.context.impl.ConfigurationServiceContext;
-import com.rackspace.papi.service.context.impl.ReportingServiceContext;
 import com.rackspace.papi.spring.SpringConfiguration;
 import net.sf.ehcache.CacheManager;
 import org.slf4j.Logger;
@@ -30,7 +29,6 @@ public class ReposeValveControllerContextManager implements ServletContextListen
       ContextAdapter ca = helper.getPowerApiContext();
 
       ca.getContext(ConfigurationServiceContext.class).contextInitialized(sce);
-      ca.getContext(ReportingServiceContext.class).contextInitialized(sce);
       ca.getContext(ReposeValveControllerContext.class).contextInitialized(sce);
 
       servletContext.setAttribute("reposeValveControllerContextManager", this);
